Load lib requirements via hooks module
