+++ lib/pluck_all/models/mongoid_extension.rb
